Move misplaced function
